% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parvo_ree_main.R
\name{calculate_rmr}
\alias{calculate_rmr}
\title{Calculate Resting Metabolic Rate}
\usage{
calculate_rmr(path, accel_path = NULL, var_limit = 15, steady_state = 15)
}
\arguments{
\item{path}{Pathname to the Parvo XLSX file.}

\item{accel_path}{Pathname to the accelerometer AGD file, Default: NULL.}

\item{var_limit}{Number to ensure changes in ventilation, oxygen consumption, and respiratory quotient do not exceed variation limit, Default: 15.}

\item{steady_state}{The number of rows at the end of the data set to consider as the steady state measurement, Default: 15}
}
\value{
Returns a tibble with the resting metabolic rate output, the length of time of steady state, and the number of the last observations used (maximum: 5).
}
\description{
Read in Parvo or Cosmed data and accelerometer data (if available) to calculate the estimated resting metabolic rate.
}
\details{
Read in Parvo or Cosmed data and accelerometer data (if available) to calculate the estimated resting metabolic rate.
}
\seealso{
\code{\link[dplyr]{filter}}
\code{\link[dplyr]{group_by}}
\code{\link[dplyr]{lead-lag}}
\code{\link[dplyr]{mutate}}
\code{\link[dplyr]{rename}}
\code{\link[dplyr]{select}}
\code{\link[dplyr]{summarise_all}}
\code{\link[lubridate]{round_date}}
\code{\link[MoveKC]{read_agd}}
}
