% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parvo_ree_main.R
\name{calculate_rmr_cv}
\alias{calculate_rmr_cv}
\title{Calculate Resting Metabolic Rate using Coefficient of Variation}
\usage{
calculate_rmr_cv(path, accel_path = NULL)
}
\arguments{
\item{path}{Pathname to the REE XLSX file.}

\item{accel_path}{Pathname to the accelerometer AGD file, Default: NULL., Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Calculate Resting Metabolic Rate using Coefficient of Variation
}
\details{
Calculate Resting Metabolic Rate using Coefficient of Variation. Currently,
it is set up to calculate coefficient of variation over 10 observations (e.g., 2.5 minutes
at 15 second epochs, 5 minutes at 30 second epochs or 10 minutes at 1 minute epochs). This
may be adjustable in the future.
}
