% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract Data}
\usage{
extract_data(path, ...)
}
\arguments{
\item{path}{Path name to the Parvo or Cosmed .xlsx files}

\item{...}{Additional arguments passed to extract_data.parvo or extract_data.cosmed.
See details for more information.}
}
\value{
Returns a data set with the Parvo or Cosmed data.
}
\description{
Extracts data from Parvo or Cosmed .xlsx files.
}
\details{
Extracts data from the Parvo or Cosmed .xlsx file. Additional arguments
can be passed to extract_data.parvo or extract_data.cosmed.
Currently, these functions only take a time_breaks argument which will return the
averages of the numeric variables at the specified time interval. Time values may include
any value that can be passed to the unit argument in the \code{\link[lubridate]{floor_date}} function
(e.g., 5 sec, 1 min, etc.). The default is set to return data at a '1 sec' interval to keep the data
the same as the input .xlsx file.
}
\seealso{
\code{\link[readxl]{read_excel}}
\code{\link[lubridate]{as_date}}
\code{\link[dplyr]{select}}, \code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise_all}}
}
