% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_meta.R
\name{extract_meta}
\alias{extract_meta}
\title{Extract Meta Data}
\usage{
extract_meta(path)
}
\arguments{
\item{path}{Path name to the .xlsx file.}
}
\value{
Returns a vector with the stored meta data from the .xlsx file.
}
\description{
Extracts the meta data from Parvo or Cosmed .xlsx files
}
\details{
Extracts the meta data from Parvo or Cosmed .xlsx files
}
