% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parvo_extract_data.R
\name{parvo_extract_data}
\alias{parvo_extract_data}
\title{Parvo Extract Data}
\usage{
parvo_extract_data(parvo.path, ree = FALSE, aee = FALSE, time.breaks = "1 sec")
}
\arguments{
\item{parvo.path}{Pathname to the Parvo XLSX file.}

\item{ree}{Logical value indicating whether you are extracting data from a resting energy expenditure measurement, Default: FALSE}

\item{aee}{Logical value indicating whether you are extracting data from a activity energy expenditure measurement, Default: FALSE}

\item{time.breaks}{Time value (e.g., 5 sec, 1 min) to indicate the interval for which the data should be aggregated, Default: '1 sec'}
}
\value{
Returns a data set with the Parvo data.
}
\description{
Extracts the Parvo data from the XLSX file.
}
\details{
Copyright © 2023 University of Kansas. All rights reserved.

Creative Commons Attribution NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0)

Extracts the Parvo data from the XLSX file.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
\code{\link[lubridate]{as_date}}
\code{\link[dplyr]{select}}, \code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise_all}}
}
