% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parvo_extract_meta.R
\name{parvo_extract_meta}
\alias{parvo_extract_meta}
\title{Parvo Extract Meta Data}
\usage{
parvo_extract_meta(parvo.path)
}
\arguments{
\item{parvo.path}{Pathname to the Parvo XLSX file.}
}
\value{
Returns a vector with the stored meta data from the Parvo XLSX file.
}
\description{
Extracts the Parvo meta data from the top of the XLSX file.
}
\details{
Copyright © 2023 University of Kansas. All rights reserved.

Creative Commons Attribution NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0)

Extracts the Parvo meta data from the top of the XLSX file.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
\code{\link[stringr]{str_replace}}
}
