% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parvoR.R
\docType{package}
\name{parvoR}
\alias{parvoR}
\title{parvoR: Useful functions to interact with data from the ParvoMedics TrueOne 2400.}
\description{
The parvoR package provides several important functions:

\code{\link{parvo_get_oues}}

\code{\link{parvo.extract.data}}

\code{\link{parvo.extract.meta}}

\code{\link{parvo.ree.main}}

\code{\link{parvo.aee.final4}}
}
