% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parvoR.R
\docType{package}
\name{parvoR}
\alias{parvoR}
\title{parvoR: Useful functions to interact with data from the ParvoMedics TrueOne 2400.}
\description{
The parvoR package provides several important functions:

\code{\link{parvo_calculate_oues}}

\code{\link{parvo_extract_data}}

\code{\link{parvo_extract_meta}}

\code{\link{parvo_ree_main}}

\code{\link{parvo_aee_final4}}
}
